refactor: optimize dictionary operations and response handling
- Improved the `extract_messages` function to utilize direct key lookups, enhancing performance.
- Updated section titles for clarity in documentation.
- Consolidated reask logic into `handle_reask_kwargs` function, removing the obsolete `reask.py` file to streamline the codebase.
- Enhanced imports in `process_response.py` for better organization and readability.
